requirements.txt
# Core AI & LangGraph
langchain
langgraph
langchain-groq

# Environment config
python-dotenv

# FastAPI WebSocket backend
fastapi
uvicorn

# (Optional) For local dev WebSocket testing via Python
websockets


# Core Packages
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Database
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver

# Facebook SDK
facebook-business==19.0.0
requests==2.31.0

# Security
cryptography==42.0.4
fernet==1.0.1

# Async Utilities
httpx==0.26.0
asyncio==3.4.3

# Development
pytest==8.0.2
pytest-asyncio==0.23.5
python-jose==3.3.0  # JWT for testing